Fix miri issues with bbbul
As we are building a forward linked list, by the time we modify a node to set its successor,
that node is already referenced in its predecessor.

That means that we cannot claim exclusive ownership to the node whose successor we want to set.
Instead, we're using interior mutability to the `next_node` inside of the `Node` structure, so
that we can do everything with shared references.